일정 메모 seq 순서 변경 작업중...

+++ src/main/java/com/green/project_quadruaple/trip/model/res/TripDetailRes.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class TripDetailRes {
 
     private long totalDistance;
-    private int totalDuration;
+    private long totalDuration;
     private int scheduleCnt;
     private int memoCnt;
     private List<TripDetailDto> days;
